Rename folder name in downstream_tasks
